style: #60 패키지 이동, 바뀐 이름으로 코드 적용
