+++ data/src/main/java/com/mashup/alcoholfree/data/dto/remote/response/DrinkResponse.kt
@@ -1,4 +1,4 @@
-package com.mashup.alcoholfree.data.remote.response
+package com.mashup.alcoholfree.data.dto.remote.response
 
 import com.google.gson.annotations.SerializedName
