+++ data/src/main/java/com/mashup/alcoholfree/data/repository/MeasureRepositoryImpl.kt
@@ -1,0 +1,17 @@
+package com.mashup.alcoholfree.data.repository
+
+import com.mashup.alcoholfree.data.datasource.MeasureDataSource
+import com.mashup.alcoholfree.data.model.MeasureResultDataModel
+import com.mashup.alcoholfree.domain.model.MeasureResult
+import com.mashup.alcoholfree.domain.repository.MeasureRepository
+import javax.inject.Inject
+
+class MeasureRepositoryImpl @Inject constructor(
+    private val measureDataSource: MeasureDataSource,
+) : MeasureRepository {
+    override suspend fun getMeasureResult(reportId: String): MeasureResult {
+        return MeasureResultDataModel.toDomainModel(
+            measureDataSource.getMeasureResult(reportId),
+        )
+    }
+}
