+++ domain/src/main/java/com/mashup/alcoholfree/domain/repository/LoginRepository.kt
@@ -1,4 +1,4 @@
-package com.mashup.alcoholfree.domain.repository.login
+package com.mashup.alcoholfree.domain.repository
 
 import kotlinx.coroutines.flow.Flow
