+++ domain/src/main/java/com/mashup/alcoholfree/domain/repository/MeasureRepository.kt
@@ -1,0 +1,7 @@
+package com.mashup.alcoholfree.domain.repository
+
+import com.mashup.alcoholfree.domain.model.MeasureResult
+
+interface MeasureRepository {
+    suspend fun getMeasureResult(reportId: String): MeasureResult
+}
