--- domain/src/main/java/com/mashup/alcoholfree/domain/repository/measure/MeasureRepository.kt
@@ -1,7 +1,0 @@
-package com.mashup.alcoholfree.domain.repository.measure
-
-import com.mashup.alcoholfree.domain.model.MeasureResultDomainModel
-
-interface MeasureRepository {
-    suspend fun getMeasureResult(reportId: String): MeasureResultDomainModel
-}
